requirements.txt
Django==2.2
celery==5.2.3
djangorestframework==3.13.1
requests==2.27.1
django-celery-results==2.3.0
psycopg2-binary==2.8.6
